Allow nested customizations in `BuildSystemProviderWrapper`
`BuildSystemProviderWrapper` is used in tests to wrap and customize the behavior
of `BuildSystemProvider`s. However it does allow customizations to the behavior
of a wrapped build system.

Make it possible to override instantiation of a `BuildSystemWrapper` in tests.

(cherry picked from commit a76579fd20ad721f8cee4378f014a983fa5f603c)