Do not attempt to build files we don't know
When building dependencies, if a given file is not know
we surface that to the user and refuse to build notifying
a new sync is needed.

While we will change this to trigger an auto-sync, it can
still be possible that the file is not part of a known rule
and we have to notify the user.

(cherry picked from commit 46f61d3c017eee3fa1177228943f93b0b29d265c)