Fix crash when sync not complete yet.
It's expected that the target map may be used before sync is compete, and we should not throw in that case. Returning empty is consistent with what the legacy sync target map does in this instance so should be safe.

(cherry picked from commit 3adca2013d492ef0d9b83cd9846c270d4f42b01f)