Add "GotIt" popup to the build dependencies action
As discussed with the UX team, this change adds
a pop up to educate users on what build dependencies
does. The pop up shows only once, and when dismissed
it never pops up again.

The pop up is the same used for reader mode and other
notifications in the IDE. This has the advantage
that when the reader mode pop up shows up, the
system will wait to not show both at the same time.

(cherry picked from commit e137195564594a6d34658bc9628f3f9b50b264fa)