Some readability improvements to ArtifactTrackerImplTest.
The helper method createOutputInfo obfuscated the structure of the OutputInfo being created, and doesn't scale as we add tests with more output groups in the OutputInfo. Since we have a builder for it already, use this in perference.

Also improve the builder API to make it a little easier to use in this context.

(cherry picked from commit 151e460b1ac96005b4949cdc3ce04d6f3959398f)