Prepare for 2023.3: Do not depend on MacroManager.getInstance() (#5390)
Probably caused by this change
https://github.com/JetBrains/intellij-community/commit/b3ed61706fbed5e04c88d1ab3f69160246fd87e8#diff-737d7a5e2b283fc8b2a9c7f301e689e29a83629c1a3533053b813570da2e59f4R21

Before that, MacroManager was instantiated only if it was present in LangExtensions.xml (not a thing in Bazel Plugin tests).
Now, it is instantiated always, which results in a crash in expandMacros call in tests. Hence, we need a new way to disable
macro expansion in tests.