Improvements to language class enums.
The motivation of this chamnge is to make the eixsting query-sync specific enums (Java and proto) usable outside of the `ProjectDefinition`. As such, move them both to the top level, and rename the Java variant to `QuerySyncLanguage` to avoid the name clash with the existing `LanguageClass`.

This has also surfaced issues with the two enums getting out of sync, and the conversion code not being complete. Improve java<->proto conversion logic and add test cases to ensure this cannot happen in future.

This allows the enums to be used in other contexts in a follow-up CL.

(cherry picked from commit 3a1108f34189f3ec95ba54cd437ed67e13ff7e2e)