Mitigate StartupActivity warnings (#5760)
Since these startup activities do not require indexes
we can make them DumbAware and the warning will not be shown.
Advised path to migrate them to ProjectActivity is slightly
challenging as it requires implementation of Kotlin suspend function
in Java code and intellij has issues with bazel plugin indexing in this
case. See https://youtrack.jetbrains.com/issue/IDEA-338613 for the
details.