Change restrictions check to allow select
This change makes the following changes:

Moves the check to the top rule, allowing to pass the list as
attributes and use a select on it.

Add check for negative, so if targets can be
deleted the rule will fail. This keeps
the list matching the dependencies

(cherry picked from commit bb6580d5ad3551f3e1db8852c1084903f6372977)