Add VCS state to the build result, and pass this through to `OutputInfo`.
This enables build systems that operate on a readonly snapshot of the workspace to communicate this information upwards. This will in turn enable accurate tracking of when query sync dependencies are built, and detection of when they are out of date.

(cherry picked from commit 5263deec3385e1084c9e006f430981a32ac9854c)