Handle ineffective rewinding for top-level outputs.
Similar to lost inputs, we limit the number of rewinding attempts before giving up and failing the build. Also renamed the existing test case for ineffective rewinding with lost inputs for consistency.

I spent some extra time making sure that the error message is well-formatted since it can possibly be displayed as the only error in the build. This is different from ineffective rewinding with lost inputs, in which case `ActionRewindException`'s message is not displayed to the user.

(cherry picked from commit 8480a17338cd94fb2804df933c00050d93eca2ab)